moved missing_vals func to eda utils
