+++ ai4water/eda/_main.py
@@ -14,7 +14,7 @@ except ModuleNotFoundError:
     sns = None
 
 from .utils import pac_yw, auto_corr, plot_autocorr
-from ai4water.utils.utils import _missing_vals
+from .utils import _missing_vals
 from ai4water.utils.visualizations import Plot
 from ai4water.utils.easy_mpl import bar_chart
 from ai4water.utils.utils import find_tot_plots
